Extracting IdExecutionService API from the instrument implementation
