+++ engine/runtime/src/main/java/org/enso/interpreter/service/ExecutionService.java
@@ -18,7 +18,7 @@ import java.util.UUID;
 import java.util.function.Consumer;
 import org.enso.compiler.context.ChangesetBuilder;
 import org.enso.interpreter.instrument.Endpoint;
-import org.enso.interpreter.instrument.IdExecutionInstrument;
+import org.enso.interpreter.instrument.IdExecutionService;
 import org.enso.interpreter.instrument.MethodCallsCache;
 import org.enso.interpreter.instrument.NotificationHandler;
 import org.enso.interpreter.instrument.RuntimeCache;
@@ -54,7 +54,7 @@ import org.enso.text.editing.model;
  */
 public class ExecutionService {
   private final Context context;
-  private final IdExecutionInstrument idExecutionInstrument;
+  private final IdExecutionService idExecutionInstrument;
   private final NotificationHandler.Forwarder notificationForwarder;
   private final InteropLibrary interopLibrary = InteropLibrary.getFactory().getUncached();
   private final TruffleLogger logger = TruffleLogger.getLogger(LanguageInfo.ID);
@@ -64,15 +64,15 @@ public class ExecutionService {
    * Creates a new instance of this service.
    *
    * @param context the language context to use.
-   * @param idExecutionInstrument an instance of the {@link IdExecutionInstrument} to use in the
-   *     course of executions.
+   * @param idExecutionInstrument an instance of the {@link IdExecutionService} to use in the course
+   *     of executions.
    * @param notificationForwarder a forwarder of notifications, used to communicate with the user
    * @param connectedLockManager a connected lock manager (if it is in use) that should be connected
    *     to the language server, or null
    */
   public ExecutionService(
       Context context,
-      IdExecutionInstrument idExecutionInstrument,
+      IdExecutionService idExecutionInstrument,
       NotificationHandler.Forwarder notificationForwarder,
       ConnectedLockManager connectedLockManager) {
     this.idExecutionInstrument = idExecutionInstrument;
@@ -142,9 +142,9 @@ public class ExecutionService {
       MethodCallsCache methodCallsCache,
       UpdatesSynchronizationState syncState,
       UUID nextExecutionItem,
-      Consumer<IdExecutionInstrument.ExpressionCall> funCallCallback,
-      Consumer<IdExecutionInstrument.ExpressionValue> onComputedCallback,
-      Consumer<IdExecutionInstrument.ExpressionValue> onCachedCallback,
+      Consumer<IdExecutionService.ExpressionCall> funCallCallback,
+      Consumer<IdExecutionService.ExpressionValue> onComputedCallback,
+      Consumer<IdExecutionService.ExpressionValue> onCachedCallback,
       Consumer<Exception> onExceptionalCallback)
       throws ArityException, SourceNotFoundException, UnsupportedMessageException,
           UnsupportedTypeException {
@@ -199,9 +199,9 @@ public class ExecutionService {
       MethodCallsCache methodCallsCache,
       UpdatesSynchronizationState syncState,
       UUID nextExecutionItem,
-      Consumer<IdExecutionInstrument.ExpressionCall> funCallCallback,
-      Consumer<IdExecutionInstrument.ExpressionValue> onComputedCallback,
-      Consumer<IdExecutionInstrument.ExpressionValue> onCachedCallback,
+      Consumer<IdExecutionService.ExpressionCall> funCallCallback,
+      Consumer<IdExecutionService.ExpressionValue> onComputedCallback,
+      Consumer<IdExecutionService.ExpressionValue> onCachedCallback,
       Consumer<Exception> onExceptionalCallback)
       throws ArityException, ConstructorNotFoundException, MethodNotFoundException,
           ModuleNotFoundException, UnsupportedMessageException, UnsupportedTypeException {
